fixes & movig tests to spec files
