+++ dpp/src/identity/state_transition/identity_update_transition/apply_identity_update_transition.rs
@@ -7,12 +7,11 @@ use crate::{
 
 use super::identity_update_transition::IdentityUpdateTransition;
 
-/// Appy Identity State Transition
-pub async fn apply_identity_update_transition<SR>(
+/// Apply Identity Update state transition
+pub async fn apply_identity_update_transition(
     state_repository: &impl StateRepositoryLike,
     state_transition: IdentityUpdateTransition,
 ) -> Result<(), ProtocolError> {
-    // let identity_id = state_transition.get_identity_id();
     let maybe_identity: Option<Identity> = state_repository
         .fetch_identity(state_transition.get_identity_id())
         .await?;
@@ -57,4 +56,60 @@ fn identity_not_found_error(identity_id: Identifier) -> ProtocolError {
     ProtocolError::AbstractConsensusError(Box::new(ConsensusError::BasicError(Box::new(
         BasicError::IdentityNotFoundError { identity_id },
     ))))
+}
+
+#[cfg(test)]
+mod test {
+    use crate::{
+        identity::state_transition::identity_update_transition::identity_update_transition::IdentityUpdateTransition,
+        state_repository::MockStateRepositoryLike,
+        tests::fixtures::{get_identity_update_transition_fixture, identity_fixture},
+    };
+
+    use super::apply_identity_update_transition;
+
+    struct TestData {
+        state_transition: IdentityUpdateTransition,
+        state_repository_mock: MockStateRepositoryLike,
+    }
+
+    fn setup_test() -> TestData {
+        let mut state_transition = get_identity_update_transition_fixture();
+        state_transition.set_revision(state_transition.get_revision() + 1);
+
+        let identity = identity_fixture();
+
+        let mut state_repository_mock = MockStateRepositoryLike::new();
+        let identity_to_return = identity.clone();
+        state_repository_mock
+            .expect_fetch_identity()
+            .returning(move |_| Ok(Some(identity_to_return.clone())));
+
+        TestData {
+            state_transition,
+            state_repository_mock,
+        }
+    }
+
+    #[tokio::test]
+    async fn should_add_public_key() {
+        let TestData {
+            mut state_transition,
+            mut state_repository_mock,
+        } = setup_test();
+
+        state_transition.set_public_keys_disabled_at(None);
+        state_transition.set_public_key_ids_to_disable(vec![]);
+        state_repository_mock
+            .expect_store_identity_public_key_hashes()
+            .returning(|_, _| Ok(()));
+        state_repository_mock
+            .expect_update_identity()
+            .returning(|_| Ok(()));
+
+        let result =
+            apply_identity_update_transition(&state_repository_mock, state_transition).await;
+
+        assert!(result.is_ok());
+    }
 }
