+++ dpp/src/identity/validation/mod.rs
@@ -1,9 +1,9 @@
 pub use identity_existence_validator::*;
 pub use identity_validator::*;
-pub use public_keys_in_identity_transition_validator::*;
+pub use required_purpose_and_security_level_validator::*;
 pub use public_keys_validator::*;
 
 mod identity_existence_validator;
 mod identity_validator;
-mod public_keys_in_identity_transition_validator;
+mod required_purpose_and_security_level_validator;
 mod public_keys_validator;
