+++ dpp/src/identity/validation/public_keys_validator.rs
@@ -10,6 +10,7 @@ use crate::errors::consensus::basic::identity::{
     InvalidIdentityPublicKeyDataError, InvalidIdentityPublicKeySecurityLevelError,
 };
 use crate::identity::{IdentityPublicKey, KeyType, ALLOWED_SECURITY_LEVELS};
+use crate::util::json_value::JsonValueExt;
 use crate::validation::{JsonSchemaValidator, ValidationResult};
 use crate::{DashPlatformProtocolInitError, NonConsensusError, PublicKeyValidationError};
 
@@ -133,9 +134,44 @@ impl TPublicKeysValidator for PublicKeysValidator {
     }
 }
 
+pub struct PublicKeysValidatorOptions {
+    pub must_be_enabled: bool,
+}
+
 impl PublicKeysValidator {
     pub fn new() -> Result<Self, DashPlatformProtocolInitError> {
         let public_key_schema_validator = JsonSchemaValidator::new(PUBLIC_KEY_SCHEMA.clone())?;
+
+        let public_keys_validator = Self {
+            public_key_schema_validator,
+        };
+
+        Ok(public_keys_validator)
+    }
+
+    pub fn new_with_options(
+        options: PublicKeysValidatorOptions,
+    ) -> Result<Self, DashPlatformProtocolInitError> {
+        let schema_for_validation = match options.must_be_enabled {
+            true => {
+                let mut schema = PUBLIC_KEY_SCHEMA.clone();
+                schema
+                    .get_mut("properties")
+                    .ok_or(DashPlatformProtocolInitError::InvalidSchemaError(
+                        "'properties' is not found",
+                    ))?
+                    .remove("disabledAt")
+                    .map_err(|_| {
+                        DashPlatformProtocolInitError::InvalidSchemaError(
+                            "'properties.disabledAt' not found",
+                        )
+                    })?;
+                schema
+            }
+            false => PUBLIC_KEY_SCHEMA.clone(),
+        };
+
+        let public_key_schema_validator = JsonSchemaValidator::new(schema_for_validation)?;
 
         let public_keys_validator = Self {
             public_key_schema_validator,
