+++ dpp/src/tests/identity/state_transition/identity_update_transition/mod.rs
@@ -1,0 +1,4 @@
+mod apply_identity_udpate_transition_spec;
+mod validate_identity_update_transition_basic_spec;
+mod validate_identity_update_transition_state_spec;
+mod validate_public_keys;
