+++ dpp/src/tests/identity/state_transition/mod.rs
@@ -2,3 +2,4 @@ pub mod asset_lock;
 mod identity_create_transition;
 mod identity_credit_withdrawal_transition;
 mod identity_topup_transition;
+mod identity_update_transition;
