+++ dpp/src/tests/identity/validation/mod.rs
@@ -1,2 +1,3 @@
 mod identity_validator_spec;
 mod public_keys_validator_spec;
+mod required_purpose_and_security_level_validator_spec;
